requirements.txt
streamlit==1.42.2
peewee==3.17.9
groq==0.18.0
python-dotenv==1.0.1